\name{hapassoc-internal}
\title{Internal hapassoc functions}
\alias{EMnull}
\alias{EMvar}
\alias{IPhi}
\alias{IPhiGamma}
\alias{IPhiGaussian}
\alias{SPhi}
\alias{SPhiGamma}
\alias{SPhiGaussian}
\alias{codeHaploDM}
\alias{getHaplos}
\alias{getPhenos}
\alias{handleMissings}
\alias{isHetero}
\alias{isIn}
\alias{isMissing}
\alias{isMultiHetero}
\alias{makeHaploLab}
\alias{makeHaploLabN}
\alias{mlPhi}
\alias{mlPhiGamma}
\alias{momentPhiGamma}
\alias{pYgivenX}
\alias{RecodeHaplos}
\description{Internal hapassoc functions.}
\details{These are not to be called by the user or are undocumented.}
\keyword{internal}
