\name{hypoDat}
\docType{data}
\alias{hypoDat}
\title{Simulated data for a hypothetical binary trait}
\description{
Simulated binary trait data used to illustrate the hapassoc package.
}
\usage{data(hypoDat)}
\format{Matrix with columns:\
\tabular{rlll}{
[,1] \tab affected \tab numeric \tab affection status (1=yes, 0=no) \cr
[,3] \tab attr \tab numeric \tab simulated quantitative attribute \cr
[,5] \tab M1.1 \tab numeric \tab the first allele of hypothetical SNP M1 \cr
[,6] \tab M1.2 \tab numeric \tab the second allele of hypothetical SNP M1 \cr
[,5] \tab M2.1 \tab numeric \tab the first allele of hypothetical SNP M2 \cr
[,6] \tab M2.2 \tab numeric \tab the second allele of hypothetical SNP M2 \cr
[,7] \tab M3.1 \tab numeric \tab the first allele of hypothetical SNP M3 \cr
[,8] \tab M3.2 \tab numeric \tab the second allele of hypotetical SNP  M3 }
}
\keyword{datasets}
