\name{hypoDatGeno}
\docType{data}
\alias{hypoDatGeno}
\title{Simulated data for a hypothetical genetic SNPs}
\description{
Simulated genetic SNPs data used to illustrate the hapassoc package.
}
\usage{data(hypoDatGeno)}
\format{Matrix with columns:\
\tabular{rlll}{
[,1] \tab affected \tab numeric \tab affection status (1=yes, 0=no) \cr
[,2] \tab attr \tab numeric \tab simulated quantitative attribute \cr
[,3] \tab M1 \tab numeric \tab hypothetical SNP M1 \cr
[,4] \tab M2 \tab numeric \tab hypothetical SNP M2 \cr
[,5] \tab M3 \tab numeric \tab hypothetical SNP M3 \cr  }
}
\keyword{datasets}
