\name{hypoDat}
\docType{data}
\alias{hypoDat}
\title{Simulated data for a hypothetical binary trait}
\description{
Simulated binary trait data used to illustrate the hapassoc package.
}
\usage{data(hypoDat)}
\format{Matrix with columns:\
\tabular{rlll}{
[,1] \tab affected \tab numeric \tab affection status (1=yes, 0=no) \cr
[,2] \tab gender \tab numeric \tab  coded as 0 for female and 1 for male \cr
[,3] \tab age \tab numeric \tab in years \cr
[,4] \tab BMI \tab numeric \tab Body Mass Index \cr
[,5] \tab M90.1 \tab numeric \tab the first allele of hypothetical SNP M90 \cr
[,6] \tab M90.2 \tab numeric \tab the second allele of hypothetical SNP M90 \cr
[,7] \tab M91.1 \tab numeric \tab the first allele of hypothetical SNP M91 \cr
[,8] \tab M91.2 \tab numeric \tab the second allele of hypotetical SNP  M91 }
}
\keyword{datasets}
